+++ tests/components/hassio/__init__.py
@@ -1,1 +1,3 @@
 """Tests for Hass.io"""
+
+API_PASSWORD = 'pass1234'
