+++ tests/components/hassio/test_handler.py
@@ -3,11 +3,7 @@ import asyncio
 import os
 from unittest.mock import patch
 
-import pytest
-
 from homeassistant.setup import async_setup_component
-
-API_PASSWORD = 'pass1234'
 
 
 @asyncio.coroutine
